projectBadges is added and deleted
Joining and leaving pojects happens through userservice as well as adding members by project leaders. When joining, the key is also added to projectBadges map of the user. When leaving, it is deleted if its value is 0. Details fragment updates the member list when the user joins or leaves the project.